+++ crates/threshold-signature-server/src/helpers/app_state.rs
@@ -19,12 +19,14 @@ use crate::{
     launch::Configuration,
     signing_client::ListenerState,
 };
-use entropy_kvdb::kv_manager::KvManager;
-use entropy_shared::X25519PublicKey;
+use entropy_kvdb::kv_manager::{helpers::serialize as key_serialize, KvManager};
+use entropy_protocol::KeyShareWithAuxInfo;
+use entropy_shared::{X25519PublicKey, NETWORK_PARENT_KEY, NEXT_NETWORK_PARENT_KEY};
 use serde::{Deserialize, Serialize};
 use sp_core::{crypto::AccountId32, sr25519, Pair};
 use std::{
     collections::HashMap,
+    path::PathBuf,
     sync::{Arc, RwLock},
 };
 use subxt::{
@@ -96,17 +98,24 @@ pub struct Cache {
     pub attestation_nonces: Arc<RwLock<HashMap<X25519PublicKey, [u8; 32]>>>,
     /// Collection of block numbers to store
     pub block_numbers: Arc<BlockNumbers>,
+    /// The network keyshare, if we have one
+    network_key_share: Arc<RwLock<Option<KeyShareWithAuxInfo>>>,
+    /// The next network keyshare, stored during reshare confirmation
+    next_network_key_share: Arc<RwLock<Option<KeyShareWithAuxInfo>>>,
 }
 
 impl Default for Cache {
     fn default() -> Self {
-        Self::new()
+        Self::new(None, None)
     }
 }
 
 impl Cache {
     /// Setup new Cache
-    pub fn new() -> Self {
+    pub fn new(
+        network_key_share: Option<KeyShareWithAuxInfo>,
+        next_network_key_share: Option<KeyShareWithAuxInfo>,
+    ) -> Self {
         Self {
             listener_state: ListenerState::default(),
             tss_state: Arc::new(RwLock::new(TssState::new())),
@@ -114,6 +123,8 @@ impl Cache {
             encryption_key_backup_provider: Default::default(),
             attestation_nonces: Default::default(),
             block_numbers: Default::default(),
+            network_key_share: Arc::new(RwLock::new(network_key_share)),
+            next_network_key_share: Arc::new(RwLock::new(next_network_key_share)),
         }
     }
     /// Returns true if all prerequisite checks have passed.
@@ -261,6 +272,7 @@ impl Cache {
             BlockNumberFields::ProactiveRefresh => self.block_numbers.proactive_refresh.clone(),
         }
     }
+
     /// Gets the list of peers who haven't yet subscribed to us for this particular session.
     pub fn unsubscribed_peers(
         &self,
@@ -272,6 +284,44 @@ impl Cache {
                 session_id,
             ))
         })
+    }
+
+    fn read_network_key_share(&self) -> Result<Option<KeyShareWithAuxInfo>, AppStateError> {
+        let key_share =
+            self.network_key_share.read().map_err(|e| AppStateError::PosionError(e.to_string()))?;
+        Ok(key_share.clone())
+    }
+
+    fn write_network_key_share(
+        &self,
+        updated_key_share: Option<KeyShareWithAuxInfo>,
+    ) -> Result<(), AppStateError> {
+        let mut key_share = self
+            .network_key_share
+            .write()
+            .map_err(|e| AppStateError::PosionError(e.to_string()))?;
+        *key_share = updated_key_share;
+        Ok(())
+    }
+
+    fn read_next_network_key_share(&self) -> Result<Option<KeyShareWithAuxInfo>, AppStateError> {
+        let key_share = self
+            .next_network_key_share
+            .read()
+            .map_err(|e| AppStateError::PosionError(e.to_string()))?;
+        Ok(key_share.clone())
+    }
+
+    fn write_next_network_key_share(
+        &self,
+        updated_key_share: Option<KeyShareWithAuxInfo>,
+    ) -> Result<(), AppStateError> {
+        let mut key_share = self
+            .next_network_key_share
+            .write()
+            .map_err(|e| AppStateError::PosionError(e.to_string()))?;
+        *key_share = updated_key_share;
+        Ok(())
     }
 }
 
@@ -285,21 +335,42 @@ pub struct AppState {
     /// Configuation containing the chain endpoint
     pub configuration: Configuration,
     /// Key-value store
-    pub kv_store: KvManager,
+    kv_store: KvManager,
     /// Global cache for TSS server
     pub cache: Cache,
 }
 
 impl AppState {
     /// Setup AppState with given secret keys
-    pub fn new(
+    pub async fn new(
         configuration: Configuration,
         kv_store: KvManager,
         pair: sr25519::Pair,
         x25519_secret: StaticSecret,
     ) -> Self {
-        Self { pair, x25519_secret, configuration, kv_store, cache: Cache::default() }
+        // Read the network keyshare from the kv_store on startup - this is the only point at which
+        // we use it as a source of truth, as it is vulnerable to rollback attacks
+        let network_key_share: Option<KeyShareWithAuxInfo> = if let Ok(key_share_bytes) =
+            kv_store.kv().get(&hex::encode(NETWORK_PARENT_KEY)).await
+        {
+            entropy_kvdb::kv_manager::helpers::deserialize(&key_share_bytes)
+        } else {
+            None
+        };
+
+        let next_network_key_share: Option<KeyShareWithAuxInfo> = if let Ok(next_key_share_bytes) =
+            kv_store.kv().get(&hex::encode(NEXT_NETWORK_PARENT_KEY)).await
+        {
+            entropy_kvdb::kv_manager::helpers::deserialize(&next_key_share_bytes)
+        } else {
+            None
+        };
+
+        let cache = Cache::new(network_key_share, next_network_key_share);
+
+        Self { pair, x25519_secret, configuration, kv_store, cache }
     }
+
     /// Convenience function to get chain api and rpc
     pub async fn get_api_rpc(
         &self,
@@ -329,6 +400,64 @@ impl AppState {
     pub fn x25519_public_key(&self) -> [u8; 32] {
         x25519_dalek::PublicKey::from(&self.x25519_secret).to_bytes()
     }
+
+    pub fn network_key_share(&self) -> Result<Option<KeyShareWithAuxInfo>, AppStateError> {
+        self.cache.read_network_key_share()
+    }
+
+    pub fn next_network_key_share(&self) -> Result<Option<KeyShareWithAuxInfo>, AppStateError> {
+        self.cache.read_next_network_key_share()
+    }
+
+    pub async fn update_network_key_share(
+        &self,
+        updated_key_share: Option<KeyShareWithAuxInfo>,
+    ) -> Result<(), AppStateError> {
+        self.cache.write_network_key_share(updated_key_share.clone())?;
+
+        // Write to on-disk store for backup
+        self.kv_store.kv().delete(&hex::encode(NETWORK_PARENT_KEY)).await.unwrap();
+        if let Some(key_share_with_aux_info) = updated_key_share {
+            let serialized_key_share = key_serialize(&key_share_with_aux_info).unwrap();
+
+            let reservation =
+                self.kv_store.kv().reserve_key(hex::encode(NETWORK_PARENT_KEY)).await.unwrap();
+            self.kv_store.kv().put(reservation, serialized_key_share.clone()).await.unwrap();
+        }
+        Ok(())
+    }
+
+    pub async fn update_next_network_key_share(
+        &self,
+        updated_key_share: Option<KeyShareWithAuxInfo>,
+    ) -> Result<(), AppStateError> {
+        self.cache.write_next_network_key_share(updated_key_share.clone())?;
+
+        // Write to on-disk store for backup
+        self.kv_store.kv().delete(&hex::encode(NEXT_NETWORK_PARENT_KEY)).await.unwrap();
+        if let Some(key_share_with_aux_info) = updated_key_share {
+            let serialized_key_share = key_serialize(&key_share_with_aux_info).unwrap();
+
+            let reservation =
+                self.kv_store.kv().reserve_key(hex::encode(NEXT_NETWORK_PARENT_KEY)).await.unwrap();
+            self.kv_store.kv().put(reservation, serialized_key_share.clone()).await.unwrap();
+        }
+        Ok(())
+    }
+
+    pub async fn rotate_key_share(&self) -> Result<(), AppStateError> {
+        let next_key_share = self.cache.read_next_network_key_share()?;
+        if next_key_share.is_none() {
+            return Err(AppStateError::CannotRotateKeyShare);
+        }
+        self.update_network_key_share(next_key_share).await?;
+        self.update_next_network_key_share(None).await?;
+        Ok(())
+    }
+
+    pub fn storage_path(&self) -> PathBuf {
+        self.kv_store.storage_path().to_path_buf()
+    }
 }
 
 /// Errors related to app state.
@@ -340,4 +469,6 @@ pub enum AppStateError {
     SessionError(String),
     #[error("Subxt: {0}")]
     Subxt(#[from] subxt::Error),
+    #[error("No next keyshare to rotate")]
+    CannotRotateKeyShare,
 }