+++ CHANGELOG.md
@@ -19,6 +19,7 @@ At the moment this project **does not** adhere to
 - Update substrate to polkadot stable2409 ([#1387](https://github.com/entropyxyz/entropy-core/pull/1387))
 - Remove deadlines in OCW ([#1411](https://github.com/entropyxyz/entropy-core/pull/1411))
 - Bump synedrion to 0.3.0 and use manul for protocol session loop ([#1392](https://github.com/entropyxyz/entropy-core/pull/1392))
+- Only read from the entropy-tss key-value store following a reboot ([#1429](https://github.com/entropyxyz/entropy-core/pull/1429))
 
 ## [0.4.0](https://github.com/entropyxyz/entropy-core/compare/release/v0.3.0...release/v0.4.0) - 2025-03-31